Address feedback from cpurcell, part 2
This picks up the original values set by the user for variable-pitch mode and the default programming height and reuses them as the cached values for areas that should be monospaced.

The remaining font heights are now scales instead.

This can behave weirdly if the theme is loaded before fonts are changed... perhaps the fonts would be better set as a custom option?